+++ src/sql/workbench/parts/commandLine/electron-browser/commandLine.ts
