+++ src/vs/workbench/workbench.main.ts
@@ -526,4 +526,4 @@ import 'sql/workbench/parts/dashboard/browser/containers/dashboardNavSection.con
 import 'sql/workbench/parts/dashboard/browser/containers/dashboardModelViewContainer.contribution';
 import 'sql/workbench/parts/dashboard/browser/core/dashboardTab.contribution';
 
-import 'sql/workbench/parts/commandLine/electron-main/commandLine.contribution';
+import 'sql/workbench/parts/commandLine/electron-browser/commandLine.contribution';
